Added conftest and utils to tests; refactoring/tidy up
